Switch from a single matrix to a series of columns with more views
This is a notable improvement on first run through, taking only .5 seconds to find all of the data points. It can run all ~160,000 regressions in about 10 seconds.

However, after a few attempts it starts to really slow down and garbage collection time increases quickly. So there must be some type instability that is causing a problem that I am missing.